JAMES-4082 Move MimePartParser and related classes from the `opensearch` module to the `mailbox-store` module
- This change facilitates easier reuse of these classes in other modules.
